+++ backends/cuda-shared/ceed-cuda-shared-basis.c
@@ -62,7 +62,7 @@ int CeedBasisApplyTensor_Cuda_shared(CeedBasis basis, const CeedInt num_elem,
 
   // Clear v for transpose mode
   if (t_mode == CEED_TRANSPOSE) {
-    CeedInt length;
+    CeedSize length;
     ierr = CeedVectorGetLength(v, &length); CeedChkBackend(ierr);
     ierr = cudaMemset(d_v, 0, length * sizeof(CeedScalar)); CeedChkBackend(ierr);
   }