+++ backends/hip-shared/ceed-hip-shared-basis.c
@@ -123,7 +123,7 @@ int CeedBasisApplyTensor_Hip_shared(CeedBasis basis, const CeedInt num_elem,
 
   // Clear v for transpose mode
   if (t_mode == CEED_TRANSPOSE) {
-    CeedInt length;
+    CeedSize length;
     ierr = CeedVectorGetLength(v, &length); CeedChkBackend(ierr);
     ierr = hipMemset(d_v, 0, length * sizeof(CeedScalar)); CeedChkBackend(ierr);
   }